+++ visualization/app/codeCharta/ui/customConfigs/uploadCustomConfigButton/uploadCustomConfigButton.component.spec.ts
@@ -1,0 +1,9 @@
+import { render, screen } from "@testing-library/angular"
+import { UploadCustomConfigButtonComponent } from "./uploadCustomConfigButton.component"
+
+describe("uploadCustomConfigButtonComponent", () => {
+	it("should render", async () => {
+		await render(UploadCustomConfigButtonComponent)
+		expect(screen.getByRole("button")).toBeTruthy()
+	})
+})